test: Move ArrayBuffer memory pressure coverage
